+++ lib/flutter_midi_command_platform_interface.dart
@@ -36,16 +36,36 @@ abstract class MidiCommandPlatform extends PlatformInterface {
     throw UnimplementedError('get devices has not been implemented.');
   }
 
+  /// Starts bluetooth subsystem. Shows an alert requesting access rights for
+  /// bluetooth.
+  Future<void> startBluetoothCentral() async {
+    throw UnimplementedError(
+        'startBluetoothCentral() has not been implemented.');
+  }
+
+  /// Stream firing events whenever the bluetooth central state changes
+  Stream<String>? get onBluetoothStateChanged {
+    throw UnimplementedError(
+        'get onBluetoothStateChanged has not been implemented.');
+  }
+
+  /// Returns the current state of the bluetooth sub system
+  Future<String> bluetoothState() async {
+    throw UnimplementedError('bluetoothState() has not been implemented.');
+  }
+
   /// Starts scanning for BLE MIDI devices.
   ///
   /// Found devices will be included in the list returned by [devices].
   Future<void> startScanningForBluetoothDevices() async {
-    throw UnimplementedError('startScanningForBluetoothDevices() has not been implemented.');
+    throw UnimplementedError(
+        'startScanningForBluetoothDevices() has not been implemented.');
   }
 
   /// Stops scanning for BLE MIDI devices.
   void stopScanningForBluetoothDevices() {
-    throw UnimplementedError('stopScanningForBluetoothDevices() has not been implemented.');
+    throw UnimplementedError(
+        'stopScanningForBluetoothDevices() has not been implemented.');
   }
 
   /// Connects to the device.
@@ -71,13 +91,25 @@ abstract class MidiCommandPlatform extends PlatformInterface {
   }
 
   Stream<MidiPacket>? get onMidiDataReceived {
-    throw UnimplementedError('get onMidiDataReceived has not been implemented.');
+    throw UnimplementedError(
+        'get onMidiDataReceived has not been implemented.');
   }
 
   /// Stream firing events whenever a change in the MIDI setup occurs.
   ///
   /// For example, when a new BLE devices is discovered.
   Stream<String>? get onMidiSetupChanged {
-    throw UnimplementedError('get onMidiSetupChanged has not been implemented.');
+    throw UnimplementedError(
+        'get onMidiSetupChanged has not been implemented.');
+  }
+
+  /// Creates a virtual MIDI source.
+  void addVirtualDevice({String? name}) {
+    throw UnimplementedError('addVirtualDevice() has not been implemented.');
+  }
+
+  /// Removes a previously created virtual MIDI source.
+  void removeVirtualDevice({String? name}) {
+    throw UnimplementedError('removeVirtualDevice() has not been implemented.');
   }
 }