+++ lib/method_channel_midi_command.dart
@@ -1,17 +1,22 @@
 import 'dart:async';
 import 'dart:typed_data';
 import 'package:flutter/services.dart';
-import 'package:flutter_midi_command_platform_interface/midi_device.dart';
 import 'flutter_midi_command_platform_interface.dart';
 
-const MethodChannel _methodChannel = MethodChannel('plugins.invisiblewrench.com/flutter_midi_command');
-const EventChannel _rxChannel = EventChannel('plugins.invisiblewrench.com/flutter_midi_command/rx_channel');
-const EventChannel _setupChannel = EventChannel('plugins.invisiblewrench.com/flutter_midi_command/setup_channel');
+const MethodChannel _methodChannel =
+    MethodChannel('plugins.invisiblewrench.com/flutter_midi_command');
+const EventChannel _rxChannel =
+    EventChannel('plugins.invisiblewrench.com/flutter_midi_command/rx_channel');
+const EventChannel _setupChannel = EventChannel(
+    'plugins.invisiblewrench.com/flutter_midi_command/setup_channel');
+const EventChannel _bluetoothStateChannel = EventChannel(
+    'plugins.invisiblewrench.com/flutter_midi_command/bluetooth_central_state');
 
 /// An implementation of [MidiCommandPlatform] that uses method channels.
 class MethodChannelMidiCommand extends MidiCommandPlatform {
   Stream<MidiPacket>? _rxStream;
   Stream<String>? _setupStream;
+  Stream<String>? _bluetoothStateStream;
 
   /// Returns a list of found MIDI devices.
   @override
@@ -19,7 +24,8 @@ class MethodChannelMidiCommand extends MidiCommandPlatform {
     var devs = await _methodChannel.invokeMethod('getDevices');
     return devs.map<MidiDevice>((m) {
       var map = m.cast<String, Object>();
-      var dev = MidiDevice(map["id"].toString(), map["name"], map["type"], map["connected"] == "true");
+      var dev = MidiDevice(map["id"].toString(), map["name"], map["type"],
+          map["connected"] == "true");
       dev.inputPorts = _portsFromDevice(map["inputs"], MidiPortType.IN);
       dev.outputPorts = _portsFromDevice(map["outputs"], MidiPortType.OUT);
       return dev;
@@ -33,6 +39,36 @@ class MethodChannelMidiCommand extends MidiCommandPlatform {
       return MidiPort(portMap["id"] as int, type);
     });
     return ports.toList(growable: false);
+  }
+
+  /// Starts bluetooth subsystem.
+  ///
+  /// Shows an alert requesting access rights for bluetooth.
+  @override
+  Future<void> startBluetoothCentral() async {
+    try {
+      await _methodChannel.invokeMethod('startBluetoothCentral');
+    } on PlatformException catch (e) {
+      throw e.message!;
+    }
+  }
+
+  /// Stream firing events whenever a change in bluetooth central state happens
+  @override
+  Stream<String>? get onBluetoothStateChanged {
+    _bluetoothStateStream ??=
+        _bluetoothStateChannel.receiveBroadcastStream().cast<String>();
+    return _bluetoothStateStream;
+  }
+
+  /// Returns the current state of the bluetooth subsystem
+  @override
+  Future<String> bluetoothState() async {
+    try {
+      return await _methodChannel.invokeMethod('bluetoothState');
+    } on PlatformException catch (e) {
+      throw e.message!;
+    }
   }
 
   /// Starts scanning for BLE MIDI devices.
@@ -56,7 +92,8 @@ class MethodChannelMidiCommand extends MidiCommandPlatform {
   /// Connects to the device.
   @override
   void connectToDevice(MidiDevice device, {List<MidiPort>? ports}) {
-    _methodChannel.invokeMethod('connectToDevice', {"device": device.toDictionary, "ports": ports});
+    _methodChannel.invokeMethod(
+        'connectToDevice', {"device": device.toDictionary, "ports": ports});
   }
 
   /// Disconnects from the device.
@@ -77,7 +114,8 @@ class MethodChannelMidiCommand extends MidiCommandPlatform {
   @override
   void sendData(Uint8List data, {int? timestamp, String? deviceId}) {
     // print("send $data through method channel");
-    _methodChannel.invokeMethod('sendData', {"data": data, "timestamp": timestamp, "deviceId": deviceId});
+    _methodChannel.invokeMethod('sendData',
+        {"data": data, "timestamp": timestamp, "deviceId": deviceId});
   }
 
   /// Stream firing events whenever a midi package is received.
@@ -89,8 +127,10 @@ class MethodChannelMidiCommand extends MidiCommandPlatform {
     _rxStream ??= _rxChannel.receiveBroadcastStream().map<MidiPacket>((d) {
       var dd = d["device"];
       // print("device data $dd");
-      var device = MidiDevice(dd['id'], dd["name"], dd["type"], dd["connected"] ?? true);
-      return MidiPacket(Uint8List.fromList(List<int>.from(d["data"])), d["timestamp"] as int, device);
+      var device = MidiDevice(
+          dd['id'], dd["name"], dd["type"], dd["connected"] == "true");
+      return MidiPacket(Uint8List.fromList(List<int>.from(d["data"])),
+          d["timestamp"] as int, device);
     });
     return _rxStream;
   }
@@ -102,5 +142,20 @@ class MethodChannelMidiCommand extends MidiCommandPlatform {
   Stream<String>? get onMidiSetupChanged {
     _setupStream ??= _setupChannel.receiveBroadcastStream().cast<String>();
     return _setupStream;
+  }
+
+  /// Creates a virtual MIDI source
+  ///
+  /// The virtual MIDI source appears as a virtual port in other apps.
+  /// Currently only supported on iOS.
+  @override
+  void addVirtualDevice({String? name}) {
+    _methodChannel.invokeMethod('addVirtualDevice', {"name": name});
+  }
+
+  /// Removes a previously addd virtual MIDI source.
+  @override
+  void removeVirtualDevice({String? name}) {
+    _methodChannel.invokeMethod('removeVirtualDevice', {"name": name});
   }
 }