+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.3.3
-- Fixed device connection value on Android
+- Improved bluetooth state handling:
+  - Start bluetooth subsystem only when you want, not automatically
+  - Allow to retrieve bluetooth state before starting scanning
+  - Allow to observe bluetooth state (poweredOn, poweredOff, ...)
 
 ## 0.3.2
 - Fixed null warning